Update Firebase iOS SDK to 9.3.0
