--- ios/platform/FirebaseInAppMessaging.xcframework/ios-arm64_armv7/FirebaseInAppMessaging.framework/Modules/module.modulemap
@@ -1,11 +1,0 @@
-framework module FirebaseInAppMessaging {
-umbrella header "FirebaseInAppMessaging.h"
-export *
-module * { export * }
-  link framework "CoreTelephony"
-  link framework "Foundation"
-  link framework "Security"
-  link framework "SystemConfiguration"
-  link framework "UIKit"
-  link "z"
-}
