--- ios/platform/FirebaseInAppMessaging.xcframework/ios-arm64_i386_x86_64-simulator/FirebaseInAppMessaging.framework/Info.plist
@@ -1,20 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
-<plist version="1.0">
-<dict>
-	<key>CFBundleExecutable</key>
-	<string>FirebaseInAppMessaging</string>
-	<key>CFBundleIdentifier</key>
-	<string>com.firebase.Firebase-FirebaseInAppMessaging</string>
-	<key>CFBundleInfoDictionaryVersion</key>
-	<string>6.0</string>
-	<key>CFBundleName</key>
-	<string>FirebaseInAppMessaging</string>
-	<key>CFBundlePackageType</key>
-	<string>FMWK</string>
-	<key>CFBundleVersion</key>
-	<string>8.12.0</string>
-	<key>DTSDKName</key>
-	<string>iphonesimulator11.2</string>
-</dict>
-</plist>
