+++ ios/platform/FirebaseInAppMessaging.xcframework/ios-arm64_x86_64-maccatalyst/FirebaseInAppMessaging.framework/Headers/FirebaseInAppMessaging-umbrella.h
@@ -1,0 +1,19 @@
+#ifdef __OBJC__
+#import <UIKit/UIKit.h>
+#else
+#ifndef FOUNDATION_EXPORT
+#if defined(__cplusplus)
+#define FOUNDATION_EXPORT extern "C"
+#else
+#define FOUNDATION_EXPORT extern
+#endif
+#endif
+#endif
+
+#import "FirebaseInAppMessaging.h"
+#import "FIRInAppMessaging.h"
+#import "FIRInAppMessagingRendering.h"
+
+FOUNDATION_EXPORT double FirebaseInAppMessagingVersionNumber;
+FOUNDATION_EXPORT const unsigned char FirebaseInAppMessagingVersionString[];
+
