+++ ios/platform/FirebaseInAppMessaging.xcframework/ios-arm64_x86_64-maccatalyst/FirebaseInAppMessaging.framework/Modules/module.modulemap
@@ -1,5 +1,5 @@
 framework module FirebaseInAppMessaging {
-umbrella header "FirebaseInAppMessaging.h"
+umbrella header "FirebaseInAppMessaging-umbrella.h"
 export *
 module * { export * }
   link framework "CoreTelephony"
