+++ ios/platform/FirebaseInAppMessaging.xcframework/ios-arm64_x86_64-simulator/FirebaseInAppMessaging.framework/Modules/module.modulemap
@@ -1,0 +1,11 @@
+framework module FirebaseInAppMessaging {
+umbrella header "FirebaseInAppMessaging-umbrella.h"
+export *
+module * { export * }
+  link framework "CoreTelephony"
+  link framework "Foundation"
+  link framework "Security"
+  link framework "SystemConfiguration"
+  link framework "UIKit"
+  link "z"
+}
